requirements.txt
psycopg2-binary
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
dlt
apache-airflow
apache-airflow[s3]